backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
crypto_cpp_py==1.4.5
cytoolz==1.0.1
ecdsa==0.18.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
keras==2.12.0
lark==1.2.2
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow_dataclass==8.7.1
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
poseidon_py==0.1.5
propcache==0.3.1
protobuf==4.25.6
py-ecc==8.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.15.2
six==1.17.0
starknet-py==0.26.2
sympy==1.12.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
toolz==1.0.0
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.20.0
